scripts/requirements.txt
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0
openpyxl>=3.1.0,<4.0
pyyaml>=6.0,<7.0
lxml>=4.9.0,<6.0
python-docx>=1.1.0,<2.0
pdfplumber>=0.10.0,<1.0
pymupdf>=1.23.0,<2.0
python-pptx>=0.6.21,<1.0
